+++ src/common/main/kotlin/me/owdding/skyocean/utils/extensions/CollectionExtensions.kt
@@ -1,7 +1,5 @@
 package me.owdding.skyocean.utils.extensions
 
-import java.util.*
-
 inline fun <K, V> Iterable<K>.associateWithNotNull(keySelector: (K) -> V?): Map<K, V> = buildMap {
     for (element in this@associateWithNotNull) put(element, keySelector(element) ?: continue)
 }
@@ -12,4 +10,4 @@ fun <T> MutableList<T>.truncate(size: Int) {
     this.dropLast(toDrop)
 }
 
-fun <T> List<T>.copy() = LinkedList(this)
+fun <T> List<T>.copy() = ArrayList(this)
